api/requirements.txt
azure-functions
msrestazure
azure-cognitiveservices-vision-computervision
firebase
requests
msrest
openai>=1.0.0